% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des.stats.R
\name{ddes.stats}
\alias{ddes.stats}
\title{Create histogram, boxplot and numeric summary}
\usage{
ddes.stats(x)
}
\arguments{
\item{x}{numeric variable}
}
\description{
Create histogram, boxplot and numeric summary
}
